Outposts: Fix Outpost attacked event
