+++ Source/Outposts/Outpost_Attacks.cs
@@ -1,0 +1,59 @@
+﻿using System.Linq;
+using KCSG;
+using RimWorld;
+using Verse;
+
+namespace Outposts
+{
+    public partial class Outpost
+    {
+        public override MapGeneratorDef MapGeneratorDef
+        {
+            get
+            {
+                if (def.GetModExtension<CustomGenOption>() is { } cGen && (cGen.chooseFromlayouts.Count > 0 || cGen.chooseFromSettlements.Count > 0))
+                    return DefDatabase<MapGeneratorDef>.GetNamed("KCSG_WorldObject_Gen");
+                return MapGeneratorDefOf.Base_Faction;
+            }
+        }
+
+        public override void PostMapGenerate()
+        {
+            base.PostMapGenerate();
+
+            foreach (var pawn in Map.mapPawns.AllPawns.Where(p => p.RaceProps.Humanlike)) pawn.Destroy();
+
+            foreach (var occupant in occupants) GenPlace.TryPlaceThing(occupant, Map.Center, Map, ThingPlaceMode.Near);
+        }
+
+        public override bool ShouldRemoveMapNow(out bool alsoRemoveWorldObject)
+        {
+            if (!Map.mapPawns.FreeColonists.Any())
+            {
+                occupants.Clear();
+                Find.LetterStack.ReceiveLetter("Outposts.Letters.Lost.Label".Translate(), "Outposts.Letters.Lost.Text".Translate(Name), LetterDefOf.NegativeEvent);
+                alsoRemoveWorldObject = true;
+                return true;
+            }
+
+            if (!Map.mapPawns.AllPawns.Where(p => p.RaceProps.Humanlike).Any(p => p.HostileTo(Faction.OfPlayer)))
+            {
+                occupants.Clear();
+                Find.LetterStack.ReceiveLetter("Outposts.Letters.BattleWon.Label".Translate(), "Outposts.Letters.BattleWon.Text".Translate(Name), LetterDefOf.PositiveEvent,
+                    new LookTargets(Gen.YieldSingle(this)));
+                foreach (var pawn in Map.mapPawns.AllPawns.Where(p => p.Faction is {IsPlayer: true} || p.HostFaction is {IsPlayer: true}))
+                {
+                    pawn.DeSpawn();
+                    occupants.Add(pawn);
+                }
+
+                RecachePawnTraits();
+                alsoRemoveWorldObject = false;
+                return true;
+            }
+
+            alsoRemoveWorldObject = false;
+            return false;
+        }
+    }
+}
